UHF-6046: Rename menu javascript
